Try out backward-cpp with Mac extensions
